Factional Cargo Part 3: Magazines now come unloaded + Commissioner and Other Weapons in cargo (#3893)
## About The Pull Request

Some cargo packs due to discounts were priced at decimal values. This is
not good, so I fixed it. In the process, I realized magazines should
probably come empty and added empty variants. As such I have also
decreased the cost of magazines to not be absolutely absurd. I mean
seriously, who pays 500 credits (half the cost of a Commander) for a
magazine for the gun? It can't be that expensive to manufacture.

Also adds other weapons I forgot into factional cargo, namely the last
member of the SMR-80 AR family, the SAW-80. Also the CM-40, and
magazines for it (the gun is also available in ruins so this now allows
it to be fielded easier.)

Additional changes involve the lowering of the price of the WT-550 PDW
from 4000 to 3000 (you might as well buy an assault rifle at that price)
and the Solarian 'GAR' Carbine for 5000.

The GAR is equivalent in firepower to an SKM following a nerf in another
PR apparently, so it has been priced as such.

## Why It's Good For The Game

Doesn't potentially destroy our economic systems by creating a decimal
value of currency anymore. Magazines generally aren't sold in gun stores
with live ammunition, and also - they probably shouldn't be nearing the
costs of the rifles.

## Changelog

:cl:
add: Commissioner, SAW-80, GAR Carbine, and CM-40 to factional cargo
fix: Decimal cargo prices
balance: Magazines now come completely empty. As a result of this, they
have now been made cheaper.
/:cl: