small pistol tweaks 2 (#3948)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
(remade pr due to check fails)

Makes the ringneck 8 rounds
Makes the commander 12 rounds

Adjust the sprites to match

## Why It's Good For The Game

Apogee and I decided this would be better

## Changelog

:cl:
add: Touches up the ringneck and commander magazine sprites
balance: Ringneck is now 8 rounds and commander is 12 rounds.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: retlaw34 <58402542+retlaw34@users.noreply.github.com>