New Ruin: Tesla Lab (#3787)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Adds a new ruin to the iceplanet ruin pool - Tesla Lab
<details><summary>Screencaps</summary>
<p>


![image](https://github.com/user-attachments/assets/190906ae-ff05-4b76-9b7c-9d58cfb7c98b)
![StrongDMM-2024-11-16 15 39
07](https://github.com/user-attachments/assets/36dffe44-0198-4439-bd71-320d5a951e69)
![StrongDMM-2024-11-16 15 39
08](https://github.com/user-attachments/assets/3d52d5e7-10a3-4c12-b1bb-65b76206f32f)


</p>
</details> 

## Why It's Good For The Game
This is an very hard ruin filled with nefarious tricks and a decent bit
of loot. I want it to challenge people.
## Changelog

:cl:
add: CLIP has lost contact with a facility in the system. Go find it. Go
see what happened. I dare you.
/:cl:

---------

Co-authored-by: Bjarl <94164348+Bjarl@users.noreply.github.com>