Fixes skirt preferences for Inteq uniforms (#3934)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Because of how Inteq clothing was pathed, uniforms were failing to get
the subtype for skirts meaning they could not be properly selected. Ex.
expecting `.../corpsman/skirt` when it was actually
`.../skirt/corpsman`. This fixes the paths for skirts and updates them
on used maps.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Lets players actually use Inteq skirtlenecks, rather than failing to
retrieve the subtype and falling back to a regular turtleneck.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
fix: skirt preferences now work for inteq uniforms
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->