Wideband table/wall type swap feature (#3932)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

No longer will wideband table frames be nearly unobtainable. By simply
using a screwdriver on a wideband wall frame, it will be swapped with a
wideband table frame. Need a wall frame but you only have a table frame?
Using a screwdriver on that will do the reverse.


![image](https://github.com/user-attachments/assets/51ee2a57-920c-4668-8a8c-5ea4c1032f78)


![image](https://github.com/user-attachments/assets/058cb0fb-9db5-4e96-92cc-4eb22407acd2)


![image](https://github.com/user-attachments/assets/03affa6a-9838-405f-b320-cad182886425)

## Why It's Good For The Game

I think it would be nice to be able to change a wideband frame from one
that can only fit on a wall to one that can go on a table at will. It
opens the door to a bit more flexibility when modifying a ship's
internals.

## Changelog

:cl:
code: wideband frames can now be swapped between table frames and wall
frames with a screwdriver
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->