Mapping hazards and types (#3233)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->
(still plenty of room for actual hazards, mostly need sprites & ideas)
This PR implements a base obj/structure/hazard, with the goal of making
subtypes of this base hazard easily customizable using variables. The
goal is to make it as easy as possible for mappers to make and customize
hazards based on the subtypes listed below.
Most hazards are activated randomly or on contact. Random hazard effects
will constantly repeat with a delay set between random_min and
random_max. Contact hazard effects either happen when walked over (not
dense) or bumped (dense), with a short cooldown. As of now, only
radiation barrels use the attacked hazard trigger, which happens when
the hazard is, of course, attacked.
base hazard
all hazards can choose to add slowdown
hazard/atmospheric
a simple hazard that emits a chosen gas, with set mols and temp
hazard/electrical
hazards that emit sparks, stuns, or tesla arcs. The bread and butter of
this system
hazard/floor
floor hazards, such as spikes, slips, and launchers
hazard/radioactive
the old radioactive barrels, with their activation procs moved to the
base hazard. otherwise nearly the same!
hazard/slowdown
require climbing over, under, or have a stick chance
hazard/spray
sprays foam, smoke, or water.

implemented electrical hazard example:
https://youtu.be/4P-67V_US6Q

hazard example suite as of july 22. more are likely to come as I get
more recommendations.
https://youtu.be/gLLdpiM-Ofg

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->
Makes it (hopefully) easy for mappers to customize simple hazards to
their needs, and adds new kinds of hazards that don't rely on sim
mechanics nearly as heavily. This allows for hazards to be easily used
in ruins without much consideration, compared to something like
emitters.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

:cl:
add: adds ruin hazards
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->