Adds manufacturer field to ship config and Autowiki (#3930)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Adds manufacturer to `ship_config_schema.json` to note the manufacturer
of the ship class, which is currently displayed in the autowiki template
but no information is generated.

Ships missing manufacturer information:
- `independent_junker.json` Assistants made this in maints with a box of
scraps
- `solgov_inkwell.json` SolGov Lorema has no manufacturer lore
- `solgov_paracelsus.json`
- `srm_elder.json` They probably just built it themselves?
- `syndicate_panacea.json` Who knows, it's purple

The two other ungenerated fields for faction name and header color are
added in #3815.


![image](https://github.com/user-attachments/assets/78dc6433-608f-4ba6-ad49-b8ec01436e1d)
(Example with the Atoll-Class)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Fills in a field on ship wikis that previously did not have information
from the game. Mostly useful for mapping or lore reference, but still
helpful.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Wiki ship templates now show ship manufacturers
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->