Removes every "Hurr Durr" from code. (#4018)
## About The Pull Request

I got rid of bigotism. In our code. I hate TG

oh also I'm not fucking up merges

## Why It's Good For The Game

Why is it *not* good for the game?

## Changelog

:cl:
del: Changed the brain-damage moodlet's text. (and other NUFC)
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->