Adds the Banshee-class Salvage Clipper [Hardliner] (#3761)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Adds the Banshee-class Salvage Clipper to the game for the Hardliner
faction in-game as pictured below:


![Screenshot 2024-11-28
141823](https://github.com/user-attachments/assets/5d568ade-287e-4fdc-9aba-7beb2d9d4aba)

<details>
<summary>StrongDMM screenshots</summary>

Screenshot with areas
![2024-11-28 14 09
05](https://github.com/user-attachments/assets/37f2b4eb-cc52-48d1-ae34-d140e335ce13)

Screenshot without areas
![2024-11-28 14 09
03](https://github.com/user-attachments/assets/5a02c7f4-11a4-48bc-bb46-f7da2b026fa5)

</details>

_A drastic refit of an ICW-era Cybersun yacht into a salvage vessel.
While the Banshee retains most of the original exterior, its interior
houses an exosuit-based Hardliner salvage team and all the facilities
they could need. Though somewhat dated, it functions well as a sleeper
ship that could easily be mistaken for any other civilian vessel._

Description revised by @Apogee-dev

## Why It's Good For The Game

This adds a new vessel for a faction lacking in not only player
spawnable vessels, but also small ships. The only currently available
Hardliner vessel (Komodo) was too large for the average Komodo crew of
4, with either semi-important roles missing crew (Engineer, medic, or in
some cases 2IC's), or the majority of the crew being one role (Deckhand,
trooper).

## Changelog

:cl:
add: Adds the Banshee-class Salvage Clipper as a playable Hardliner
vessel
/:cl: