Rechambers the Firestorm to .44 Roumain (#3987)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

In a bid to make the Firestorm more flavorful, it is now chambered in
.44 Roumain. The magazine size and rate of fire have both been reduced
to account for this.
- Rechambered to .44 Roumain
- Rate of fire reduced to 0.3 SEC (a little less than half its previous
rate of fire)
- Magazine size reduced from 28 -> 24 and 50 -> 40

## Why It's Good For The Game

firestorm was sort of boring. Now it has a more interesting niche- kind
of like a heavy PDW, since it fires a cartridge with (relatively) armor
piercing characteristics.

## Changelog

:cl:
balance: SRM's Firestorm SMG is now chambered in .44 Roumain.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->