M20 Auto Elite (#3992)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

![handcannon](https://github.com/user-attachments/assets/bd606c37-584c-4540-ba45-984632c4c9ee)

Adds the M20 Auto Elite, a heavy pistol manufactured by Serene Outdoors
and chambered in .44 Roumain (subject to change in future ammo reworks).
Originally designed for Star City's police department after a series of
incidents proved the need for a handgun more powerful than the 9mm and
.45 pistols they had been using, the Auto Elite is a weighty slab of
steel with the capacity for ten powerful .44 Roumain rounds, capable of
punching through armor better than other handguns while still dealing
substantial damage. While it proved bulky and inconvenient for police
use, it found a niche as a sidearm for big game hunters and as a flashy
option for mercenaries and other people with something to prove.

Also adds the PO-20 Pinscher, Inteq's modification of the Auto Elite.
Being more readily available than 5.7 and 4.6mm options while also
having a heavier impact on target, Inteq has chosen the Auto Elite as a
standard sidearm for Vanguards and other ranking personnel. Rollout is
still in-progress, so Inteq ships on deployment will have to requisition
them on their own for now.

Both versions of the Auto Elite are available in cargo, with the
Pinscher being faction-locked to Inteq.

## Why It's Good For The Game

fills a niche and fleshes out Serene's catalogue a bit. Now indies have
access to a semi-auto heavy pistol that nevertheless has a bit less
punch than something like the CM-357, and Inteq has a designated officer
pistol.

## Changelog

:cl:
add: Added M20 Auto Elite and PO-20 Pinscher heavy pistols.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->